// src/Types/interfaces.tsx
// Generated by https://quicktype.io

// Data Cliente

export interface Cliente {
    success: boolean;
    message: string;
    data: Data;
}

export interface Data {
    id: number;
    imagen: null;
    nombres: string;
    apellidos: string;
    telefono: string;
    fechanacimiento: string;
    sexo: string;
    email: string;
    estado: string;
    verificado: number;
    token: string;
}


// Data Productos

export interface Productos {
    success: boolean;
    message: string;
    data: Producto[];
}

export interface Producto {
    id: number;
    nombre: string;
    stock: number;
    precio: number;
    imagen: string;
    descripcion: string;
    idsubcategoria: number;
    idempresa: number;
}
// Generated by https://quicktype.io

export interface Pedido {
    success: boolean;
    message: string;
    data: DataPedido[];
}

export interface DataPedido {
    id_producto: number;
    cantidad: number;
}

// Generated by https://quicktype.io

export interface ResponseMyPedido {
    success: boolean;
    message: string;
    data: MiPedido[];
}

export interface MiPedido {
    idpedido: number;
    total: string;
    productos: ProductoParcial[];
}

export interface PedidosAdapter {
    idpedido: number,
    title: string,
    data: ProductoParcial[]
}

export interface ProductoParcial {
    idproducto: number;
    nombre: string;
    precio: number;
}


// 
// Generated by https://quicktype.io

export interface ResponseProductCalificar {
    success: boolean;
    message: string;
    data: ProductCalificar[];
}

export interface ProductCalificar {
    Iddetalle: number;
    idPedido: number;
    idproducto: number;
    nombre: string;
    precio: number;
    imagen: string;
    descripcion: string;
    idsubcategoria: number;
    comentario: null | string;
    calificacion: null | number;
    promedio: null | number;
}

// Azure IA

// Generated by https://quicktype.io

export interface ResponseAzureSentiment {
    documents: Document[];
    errors: any[];
    modelVersion: string;
}

export interface Document {
    id: string;
    sentiment: string;
    confidenceScores: ConfidenceScores;
    sentences: Sentence[];
    warnings: any[];
}

export interface ConfidenceScores {
    positive: number;
    neutral: number;
    negative: number;
}

export interface Sentence {
    sentiment: string;
    confidenceScores: ConfidenceScores;
    offset: number;
    length: number;
    text: string;
}


//  Return IA_Azure

export interface ResultForComment {
    calificacion: string,
    promedio: number
}

// Generated by https://quicktype.io

export interface ReponseMessageProduct {
    success: boolean;
    message: string;
    points: string;
    data: DataProduct[];
}

export interface DataProduct {
    idproducto: number;
    nombre: string;
    promedio: string;
    calificacion: string;
    comentario: string;
}